fixed customers, started query builder

+++ app/Http/Controllers/CustomerController.php
@@ -19,9 +19,26 @@ class CustomerController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
         $user = Auth::user();
+        $shop = DB::table("shops")->where('user_id', $user->id);
+        $connections = DB::table("connections")->where('shop_id', $shop->id);
+        $page = $request->query("page")-1;
+        $key = $request->query("searchKey");
+        $sFor = $request->query("searchIn");
+        $order = $request->query("orderBy");
+        if($request->query("asc")){
+            $asc = "asc";
+        }
+        else{
+            $asc = "desc";
+        }
+        $sFor = $request->query("status");
+        $results = DB::table("customers")->where('shop_id', $shop->id)->where($sfor, 'like', '%'.$key.'%')->orderBy($order, $asc)->skip($page*30)->take(30);
+        foreach ($connections as $con){
+            DB::table("customers")->where('id', $con->customer_id)->where($sfor, 'like', '%'.$key.'%')->orderBy($order, $asc)->skip($page*30)->take(30);
+        }
         return $user;
     }
 
@@ -33,7 +50,21 @@ class CustomerController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        $user=User::findOrFail(Auth::user()->id);
+        $shop=Shop::findOrFail($shop->user_id);
+        $customer=new customer;
+        $customer->name = $request->input('name');
+        $customer->idCardNum = $request->input('idCardNum');
+        $customer->birthday = $request->input('birthday');
+        $customer->idCardExp = $request->input('idCardExp');
+        $customer->user_id = null;
+        $customer->shop_id = $shop->id;
+        $customer->shippingAddress = $request->input('shippingAddress');
+		$customer->billingAddress = $request->input('billingAddress');
+		$customer->mobile = $request->input('mobile');
+        $customer->email = $request->input('email');
+		$customer->save();
+        
     }
 
     /**
@@ -54,11 +85,11 @@ class CustomerController extends Controller
                 'message' => 'Az elem nem létezik!'
             ],404);
         }
+        
     }
 
     public function showNU(customer $customer)
     {
-        //
     }
 
 
@@ -71,21 +102,38 @@ class CustomerController extends Controller
      */
     public function update(Request $request)
     {
-        $customer=customer::findOrFail(Auth::user()->id);
-        $user=User::findOrFail($shop->user_id);
-        $customer->idCardNum = $request->input('idCardNum');
-        $customer->birthday = $request->input('birthday');
-        $customer->idCardExp = $request->input('idCardExp');
-        $customer->user_id = $request->input('user_id');
-        $customer->shop_id = $request->input('shop_id');
-        $customer->shippingAddress = $request->input('shippingAddress');
-		$customer->billingAddress = $request->input('billingAddress');
-		$customer->mobile = $request->input('mobile');
-		$customer->email = $request->input('email');
-        $user->email = $request->input('email');
-        $user->iban = $request->input('iban');
-        $user->save();
-		$customer->save();
+        $user=User::findOrFail(Auth::user()->id);
+        $customer=customer::where("user_id", $user->id);
+        if ($customer){
+            $customer->idCardNum = $request->input('idCardNum');
+            $customer->birthday = $request->input('birthday');
+            $customer->idCardExp = $request->input('idCardExp');
+            $customer->user_id = $request->input('user_id');
+            $customer->shop_id = $request->input('shop_id');
+            $customer->shippingAddress = $request->input('shippingAddress');
+            $customer->billingAddress = $request->input('billingAddress');
+            $customer->mobile = $request->input('mobile');
+            $customer->email = $request->input('email');
+            $user->email = $request->input('email');
+            $user->iban = $request->input('iban');
+            $user->save();
+            $customer->save();
+        }else{
+            $shop=Shop::where("user_id", $user->id);
+            $customer=customer::where("shop_id", $shop->id)->where("id", $request->input('id'));
+            $customer->idCardNum = $request->input('idCardNum');
+            $customer->birthday = $request->input('birthday');
+            $customer->idCardExp = $request->input('idCardExp');
+            $customer->user_id = $request->input('user_id');
+            $customer->shop_id = $request->input('shop_id');
+            $customer->shippingAddress = $request->input('shippingAddress');
+            $customer->billingAddress = $request->input('billingAddress');
+            $customer->mobile = $request->input('mobile');
+            $customer->email = $request->input('email');
+            $customer->save();
+        }
+        
+        
     }
 
     public function create(Request $request)
@@ -118,6 +166,7 @@ class CustomerController extends Controller
         $customer->shippingAddress = $request->input('shippingAddress');
 		$customer->billingAddress = $request->input('billingAddress');
 		$customer->mobile = $request->input('mobile');
+        $customer->email = $request->input('email');
 		$customer->save();
     }
 
@@ -131,7 +180,7 @@ class CustomerController extends Controller
     {
         $user = User::find(Auth::user()->id);
         $customer = Customer::find($customerId);
-        if ($customer->user_id == NULL){
+        if ($customer->user_id == NULL && $customer->shop_id ==$user->id){
             $customer->delete();
         }
         else if ($customer->user_id == $user->id){
